requirements.txt
attrs==24.2.0
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
flasgger==0.9.7.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mistune==3.0.2
ordered-set==4.1.0
packaging==24.1
Pygments==2.18.0
PyJWT==2.9.0
pyodbc==5.2.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.1
wrapt==1.16.0